move DefaultConfigurationManager
