+++ pac4j-saml/src/main/resources/META-INF/services/org.pac4j.saml.util.ConfigurationManager
@@ -1,1 +1,1 @@
-org.pac4j.saml.util.Configuration$DefaultConfigurationManager
+org.pac4j.saml.util.DefaultConfigurationManager
